requirements.txt
librosa
xgboost
streamlit-audiorecorder
numpy
sklearn